+++ mod.ts
@@ -8,10 +8,11 @@ import search, { Options as SearchOptions } from "./plugins/search.ts";
 import paginate, { Options as PaginateOptions } from "./plugins/paginate.ts";
 import toml, { Options as TomlOptions } from "./plugins/toml.ts";
 import yaml, { Options as YamlOptions } from "./plugins/yaml.ts";
+import { getOptionsFromCli } from "./core/utils/cli_options.ts";
+import { parseDateFromFilename } from "./core/utils/date.ts";
 
 import type { DeepPartial } from "./core/utils/object.ts";
 import type { SiteOptions } from "./core/site.ts";
-import { getOptionsFromCli } from "./core/utils/cli_options.ts";
 
 export interface PluginOptions {
   url?: UrlOptions;
@@ -49,6 +50,7 @@ export default function lume(
     .ignore("deno.jsonc")
     .ignore("deno.lock")
     .mergeKey("tags", "stringArray")
+    .parseFilename(parseDateFromFilename)
     .use(url(pluginOptions.url))
     .use(json(pluginOptions.json))
     .use(markdown(pluginOptions.markdown))