+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 
 ## [Unreleased]
 ### Added
+- New function `site.parseFilename`, to register new custom parsers to extract data from filenames.
 - New plugin `sri`.
 - Improved plugin docs with links to the online documentation.
 - New functions `nav.nextPage()` and `nav.previousPage()` for `nav` plugin.